Restructuring, adding license file, updating dependencies, adding ruff options
