+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "geneweaver-testing"
-version = "0.0.1a7"
+version = "0.0.1b0"
 description = "A library to standardize testing of GeneWeaver pacakges."
 authors = ["Alexander Berger <alexander.berger@jax.org>"]
 readme = "README.md"
@@ -13,18 +13,16 @@ packages = [
 
 [tool.poetry.dependencies]
 python = "^3.8"
-ruff = "^0.0.254"
-black = "^23.1.0"
+ruff = "^0.0.277"
+black = "^23.3.0"
 isort = "^5.12.0"
-pytest = "^7.2.2"
-mypy = "^1.0.1"
+pytest = "^7.4.0"
+mypy = "^1.4.1"
 tomli = "^2.0.1"
-pytest-cov = "^4.0.0"
+pytest-cov = "^4.1.0"
 
 [tool.ruff]
-select = [
-    "E", "F", "W", "C90", "N", "B", "ANN", "D", "I", "ERA", "PD", "NPY", "PT", "A", "W"
-]
+select = ['F', 'E', 'W', 'A', 'C90', 'N', 'B', 'ANN', 'D', 'I', 'ERA', 'PD', 'NPY', 'PT']
 
 [build-system]
 requires = ["poetry-core"]
