+++ src/geneweaver/testing/package/__init__.py
@@ -1,5 +1,6 @@
 """All tests relating to the package and its structure."""
 # ruff: noqa: F403
 from .can_import import *
+from .generic import *
 from .pyproject import *
 from .structure import *
