+++ src/geneweaver/testing/package/generic/__init__.py
@@ -1,0 +1,4 @@
+"""All non-geneweaver specific tests relating to the package and its structure."""
+# ruff: noqa: F403
+from .pyproject import *
+from .structure import *
