+++ src/geneweaver/testing/style/__init__.py
@@ -1,5 +1,4 @@
 """All tests relating to package code style."""
 # ruff: noqa: F403
 from .black import *
-from .isort import *
 from .ruff import *
